search workering

+++ src/main/java/app/persistence/dao/IDAO.java
@@ -5,7 +5,7 @@ import java.util.Optional;
 
 public interface IDAO<T> {
     T save(T dto);
-    Optional<T> findById(int id);
+    Optional<T> findById(long id);
     List<T> findAll();
     T update(T dto);
     void delete(int id);
